shellcheck: add expansion check so we don't wipe out '/*'

+++ tests/core/set_context.bats
@@ -3,7 +3,7 @@
 load '../test-helper'
 
 setup() {
-   rm -rf ${LABORATORY}/*
+   rm -rf ${LABORATORY:?}/*
 }
 
 @test 'set_context fails without context paramter (argument 1)' {
